+++ code/qcommon/vm.c
@@ -2485,12 +2485,6 @@ VM_Call(vm_t *vm, qint nargs, qint callnum, ...)
   }
 #endif
 
-  //reset syscall counter for top-level calls to detect infinite loops
-  if (vm->callLevel == 0)
-  {
-    vm->syscallCount = 0;
-  }
-
   ++vm->callLevel;
 
   //if we have a dll loaded, call it directly